clean up code + add realm.json w/ app id
