+++ examples/node/expo-app/app/components/LoadingSpinner.tsx
@@ -1,0 +1,6 @@
+
+import { Text, } from "react-native";
+
+const LoadingSpinner = () => <Text>Mock Loading Spinner</Text>;
+
+export default LoadingSpinner;
