+++ examples/node/expo-app/ios/Podfile.properties.json
@@ -1,0 +1,3 @@
+{
+  "expo.jsEngine": "jsc"
+}
