+++ examples/node/expo-app/ios/exporealmrntemplate/AppDelegate.h
@@ -1,0 +1,9 @@
+#import <Foundation/Foundation.h>
+#import <React/RCTBridgeDelegate.h>
+#import <UIKit/UIKit.h>
+
+#import <Expo/Expo.h>
+
+@interface AppDelegate : EXAppDelegateWrapper <RCTBridgeDelegate>
+
+@end
