+++ examples/node/expo-app/ios/exporealmrntemplate/Supporting/Expo.plist
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+        <key>EXUpdatesSDKVersion</key>
+        <string>YOUR-APP-SDK-VERSION-HERE</string>
+        <key>EXUpdatesURL</key>
+        <string>YOUR-APP-URL-HERE</string>
+</dict>
+</plist>
