+++ examples/node/expo-app/package.json
@@ -6,7 +6,7 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@realm/react": "^0.2.0",
+    "@realm/react": "^0.3.0",
     "expo": "^44.0.3",
     "expo-asset": "^8.4.6",
     "expo-dev-client": "~0.8.0",
