+++ examples/node/expo-app/realm.json
@@ -1,0 +1,3 @@
+{
+  "appId": "expo-app-ypamg"
+}
